api/requirements.txt
Flask==2.0.0
PyJWT==2.1.0